Obtain status of KQs and improve progress view style
The REST API now returns the status of KQs, two booleans values
for complete and correct states. This booleans are randomly
generated because the mongodb part is not ready yet. This will
have to be changed.

On the client side the status of KQs is processed and everything
is working like that status is the real one. No changes in the
client side will be required.

Improve style in progress view, and small fix in classroom view.
Add a bigger version of bootstrap with labels/badges, icons and
wells.