Implemented peer review submission assignment
Created assign_review view, this view tries to find a suitable submission
and if one is found, it will be assigned to the user.

Either if a suitable submission is found or not, the user will be redirected
to the main peer review assignment page. If a submission could not be assigned,
the user will be informed of the reason.