Add class AssetAvailability
New model class AssetAvailability added in order to simplify relation between
kwnowledgequantums and assets, and make it possible to store data such as when
an asset can be booked.

course_get_assets method was modified to handle the new schema.

Asset cache invalidation is yet to be adapted for the new schema.