Initial assets overview page
Initial Assets overview page was created, once it's finished it should show the
current reservations made by the user, show the bookable assets, and in which
kqs and in what date they can be booked, and allow to book them.

This initial version does only show the assets which are avaliable in a course,
without allowing to book them or to see the current reservations.