Assets student interface refactoring
The student interface for asset was modified in order to make it similar to the
classes interface.

The student will have an area in the right side in which to check if there are
active reservations and can also select an unit or KQ to book an asset.

Only KQ which have assets avaliable for reservation will appear on the list.

The current version is not yet operative.