Create a dialog for making new reservations in classroom view
A new modal was added to classroom template to contain the new reservation form.

More information about the assets is shown in the Asset view.
A button for launching the reservation form is shown as well during the booking
period. If the user clicks the button, the modal dialog appears showing the
information about the asset and a form for selectiong the date and the time.

Once the user has clicked the confirm button, a petition is send to
reservation_create view in order to create the reservation.