Do not check if the asset availability is the active tab
When saving, values of asset availability were only being updated when that tab
was the only one being active, to prevent the values being overwritten when the
instance was created.

Instead of that, it will be checked that the tab is not hidden, so if the dates
are changed the new values would be stored even if the asset availability is
not the active tab.