Update jQuery UI
jQuery UI was updated from 1.9.2 to 1.10.3
The installed version is a custom jQuery UI which includes the functionality
that was already present in the custom 1.9.2 (core, widget, mouse, draggable,
selectable, resizable and sortable) plus datepicker.

The old css file will not be updated but a css for datepicker has been added.