Display only times in which there are available slots
ReservationCount resource has been created, this resource can only be accessed
by requesting a list indicating the desired day and the id of the asset the
information is being requested. The resource will contain information about the
number of reservations which exist in each suitable time.

Classroom view was modified so only time when there are available slots are
available for selection in the combo box.