Denote the occupation of each day in the booking dialog
In the date selection calendar of the booking dialog, the color of the
background of each day will depend on the occupation of that day.

OccupationInformation resource will provide the needed information.

Created helper function get_occupation_for_month which will query the cache and
the database for obtaining the information.